+++ CHANGELOG.md
@@ -54,5 +54,6 @@
 ## 0.14.0
 * Enable customization of data type and sproc names
 * Even faster extraction of query parameter values
+* Fix bug that would sometimes result in truncating string values
 
 [FastMember]: https://github.com/mgravell/fast-member